--- src/main/java/org/ishugaliy/consistent/hash/annotation/Generated.java
@@ -1,10 +1,0 @@
-package org.ishugaliy.consistent.hash.annotation;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-/**
- * Used by Jacoco test coverage tool to exclude items from scan
- */
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Generated { }
