--- src/main/java/org/ishugaliy/consistent/hash/hasher/Hasher.java
@@ -1,7 +1,0 @@
-package org.ishugaliy.consistent.hash.hasher;
-
-@FunctionalInterface
-public interface Hasher {
-
-    long hash(String key, int seed);
-}
