--- src/main/java/org/ishugaliy/consistent/hash/node/Node.java
@@ -1,6 +1,0 @@
-package org.ishugaliy.consistent.hash.node;
-
-public interface Node {
-
-    String getKey();
-}
